updated Dockerfile to run the main.py script
Detailed message: Fixed the issue where logger.setup_logger() and init_tables() were not being called when the Docker container was started. Updated the Dockerfile to run the main.py script as the main command (CMD). This ensures that the logger setup and table initialization are performed correctly when running in AWS AppRunner or other containerized environments. Additionally, PYTHONPATH was set correctly in the Dockerfile to allow proper importing of the webapp package.